all updated

+++ PandemicShootout/app/src/main/java/com/example/pandemicshootout/vaccine/Available_vaccine.java
@@ -63,12 +63,10 @@ public class Available_vaccine extends AppCompatActivity {
                             JSONArray jsonArray = response.getJSONArray("sessions");
                             for(int i = 0; i< jsonArray.length(); i++){
                                 JSONObject k = (JSONObject) jsonArray.get(i);
-
-//                                JSONArray jsonObject = response.getJSONObject(i);
                                 Log.d("OBJECT",k.toString());
                                 String name = k.getString("name");
                                 String vac_name = k.getString("vaccine");
-                                long available = k.getLong("available_capacity");
+                                int available = k.getInt("available_capacity");
                                 long age_limit = k.getLong("min_age_limit");
 
                                 m.add(new VaccineView(name,available,age_limit,vac_name));
@@ -87,45 +85,10 @@ public class Available_vaccine extends AppCompatActivity {
                 });
 
 
-//        JsonArrayRequest jsonArrayRequest = new JsonArrayRequest(Request.Method.GET, url, null, new Response.Listener<JSONArray>() {
-//            @Override
-//            public void onResponse(JSONArray response) {
-//                vacclist.setAdapter(md);
-//                Log.d("RESPONSE",response.toString());
-//                response = response.get(0);
-//                try {
-//                    for(int i = 0; i< response.length(); i++){
-//                        JSONObject jsonObject = response.getJSONObject(i);
-//                        Log.d("OBJECT",jsonObject.toString());
-//                        boolean can = jsonObject.getBoolean("enabled");
-//                        if(can) {
-//                            String name = jsonObject.getString("name");
-//                            String vac_name = jsonObject.getString("vaccine");
-//                            long available = jsonObject.getLong("available_capacity");
-//                            long age_limit = jsonObject.getLong("min_age_limit");
-//
-//                            m.add(new VaccineView(name,available,age_limit,vac_name));
-//                        }
-//                    }
-//                } catch (Exception e) {
-//                    Toast.makeText(Available_vaccine.this,e.getMessage(),Toast.LENGTH_SHORT).show();
-//                }
-//            }
-//        }, new Response.ErrorListener() {
-//            @Override
-//            public void onErrorResponse(VolleyError error) {
-//                Toast.makeText(Available_vaccine.this,error.getMessage(),Toast.LENGTH_SHORT).show();
-//            }
-//        });
         queue.add(jsonObjectRequest);
 
 
         vacclist.setAdapter(md);
-//        m.add(new MedicineView(30,4,"Covaxin","CHC1"));
-//        m.add(new MedicineView(30,4,"Covishield","Narouli1"));
-//        m.add(new MedicineView(30,4,"Solvin cold","Mystore1"));
-//        m.add(new MedicineView(30,4,"Solvin cold","Mystore1"));
-//        m.add(new MedicineView(30,4,"Solvin cold","Mystore1"));
 
     }
 }
